add DBConnectorRedis
