+++ tests/unit/test_db_connector.py
@@ -1,6 +1,6 @@
 import pytest
 import json
-from truck_delivery_atge.db_connector import DBConnector
+from truck_delivery_atge.db_connector_redis import DBConnectorRedis
 
 
 @pytest.fixture
@@ -11,7 +11,7 @@ def fixture_db_connector(mocker):
     mocker.patch("redis.Redis.scan_iter").return_value = "123"
     mocker.patch("redis.Redis.mget").return_value = [json.dumps(object_to_save)]
     mocker.patch("redis.Redis.delete").return_value = 1
-    db_conn = DBConnector.instance()
+    db_conn = DBConnectorRedis.instance()
     return db_conn
 
 
@@ -38,7 +38,7 @@ def test_get_by_id_mock_as_context_manager(mocker):
     mocker.patch("redis.Redis.get").return_value = json.dumps(object_to_save)
 
     result = None
-    with DBConnector.instance() as db_connector:
+    with DBConnectorRedis.instance() as db_connector:
         db_connector.save(id_test, object_to_save)
         result = db_connector.get_by_id(id_test)
     assert result
@@ -54,6 +54,6 @@ def test_get_all_mock(fixture_db_connector):
 def delete_by_id_mock(fixture_db_connector):
     id_test = "123"
     object_to_save = {"123": {"a": "a"}}
-    fixture_db_connector.save(id_test)
+    fixture_db_connector.save(id_test, object_to_save)
     result = fixture_db_connector.delete(id_test)
     assert result == 1
